Exporting: Move UniversalExporter to OS trait extension
The UniversalExporter only has public methods that work on all OS's we
support. However, internally, it needs to handle per-OS differences.

Move the UniversalExporter struct to our established OS trait extension
model.

Have each OS implement how to create and setup a session and parse until
the closure invocation returns true.
